requirements.txt
aiohttp>=3.5.4
async-timeout>=3.0.1
attrs>=19.1.0
certifi>=2019.3.9
chardet>=3.0.4
Click>=7.0
discord.py>=2.0.0
idna>=2.8
itsdangerous>=1.1.0
Jinja2>=2.10.1
MarkupSafe>=1.1.1
multidict>=4.5.2
quart>=0.17.0
requests>=2.21.0
urllib3>=1.24.3
websockets>=6.0
Werkzeug>=0.15.3
yarl>=1.3.0
